Move {Format,Tree}Ops to util package.
